test: keep catalog coverage in integration lane
